+++ PlotInterface/AutoPlotHazard.py
@@ -40,7 +40,6 @@ from PlotInterface.maps import saveHazardMap
 from PlotInterface.curves import saveHazardCurve
 
 import sqlite3
-import unicodedata
 
 log = logging.getLogger(__name__)
 log.addHandler(logging.NullHandler())
@@ -225,7 +224,7 @@ class AutoPlotHazard(object):
             log.debug("Saving hazard curve for %s to %s"%(name, filename))
             wspd = ncobj.variables['wspd'][:, j, i]
 
-            recs = database.locationRecords(self.db, pID)
+            recs = database.queries.locationRecords(self.db, pID)
             data = np.zeros(int(self.numsimulations * 365.25))
             if len(recs) > 0:
                 data[-len(recs):] = recs['wspd']
